# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODBUS Consumer APIs",
        "description": "L5 Swagger OpenApi description for ODBUS Consumer APIs",
        "version": "0.1"
    },
    "paths": {
...
        "/api/Register": {
            "post": {
                "tags": [
                    "Register API"
                ],
                "summary": "send OTP to user for registration",
                "description": "user detatils saved and otp generated",
                "operationId": "c5b797bcf029bf32893c2fbfa5ba43d3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "mobile number of user",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "created_by",
                        "in": "query",
                        "description": "created by",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "otp generated"
                    },
                    "206": {
                        "description": "not a valid credential"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/VerifyOtp": {
            "post": {
                "tags": [
                    "Verify Otp"
                ],
                "summary": "otp verification",
                "description": "otp verification",
                "operationId": "c1d15a602b74f91486447ee7b513f6cc",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "user Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "otp",
                        "in": "query",
                        "description": "otp sent to user",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registered successfully"
                    },
                    "206": {
                        "description": "otp not provided"
                    },
                    "406": {
                        "description": "Invalid otp"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/Login": {
            "post": {
                "tags": [
                    "Login API"
                ],
                "summary": "user login using phone or email",
                "description": "user login using phone or email",
                "operationId": "4cc8efd8726a2d1ae178f987ca09cbd3",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone of user",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "otp generated"
                    },
                    "206": {
                        "description": "not a valid credential"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/UserProfile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get user details",
                "operationId": "82e9f7bb9d7d088c86d0661daf575808",
                "responses": {
                    "200": {
                        "description": "Authorized User details"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/updateProfile": {
            "post": {
                "tags": [
                    "update User Profile details"
                ],
                "summary": "Update User Profile details",
                "operationId": "bca8ec1334c185327dfa7313a2b4d925",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "token",
                        "in": "query",
                        "description": "User Token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "E7CDunOAhJ"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "user name",
                        "schema": {
                            "type": "string",
                            "example": "SAM"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "user email",
                        "schema": {
                            "type": "string",
                            "example": "abcd@gmail.com"
                        }
                    },
                    {
                        "name": "pincode",
                        "in": "query",
                        "description": "pincode",
                        "schema": {
                            "type": "number",
                            "example": 122345
                        }
                    },
                    {
                        "name": "street",
                        "in": "query",
                        "description": "street",
                        "schema": {
                            "type": "string",
                            "example": "St marry road"
                        }
                    },
                    {
                        "name": "district",
                        "in": "query",
                        "description": "district",
                        "schema": {
                            "type": "string",
                            "example": "Bangalore"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "address",
                        "schema": {
                            "type": "string",
                            "example": "Bangalore"
                        }
                    },
                    {
                        "name": "profile_image",
                        "in": "query",
                        "description": "profile image",
                        "schema": {
                            "type": "string",
                            "example": "profile.png"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update User Profile"
                    },
                    "206": {
                        "description": "validation error"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "AUTH0",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}